Update user and deductions models

+++ app/models/deductions_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, Numeric
 from app.database.database import Base
 
 
@@ -10,4 +10,4 @@ class DeductionType(Base):
     __tablename__ = "deduction_types"
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, unique=True, nullable=False)  # e.g., "Unmade Bed"
-    default_amount = Column(Float, nullable=False)      # e.g., 0.50
+    default_amount = Column(Numeric(10, 2), nullable=False)      # e.g., 0.50
